refactor(eval-runtime): 按捕获编译与结果复用拆分评测入口
