+++ src/eval-runtime/evaluation/ordering.ts
@@ -1,0 +1,5 @@
+
+
+export function compareStrings(left: string, right: string): number {
+  return left < right ? -1 : left > right ? 1 : 0;
+}
